buildable

+++ packages/protocol/contracts-0.8/common/linkedlists/AddressLinkedList.sol
@@ -97,10 +97,10 @@ library AddressLinkedList {
   }
 
   function toBytes(address a) public pure returns (bytes32) {
-    return bytes32(uint256(a) << 96);
+    return bytes32(uint256(uint160(a)) << 96);
   }
 
   function toAddress(bytes32 b) public pure returns (address) {
-    return address(uint256(b) >> 96);
+    return address(uint160(uint256(b) >> 96));
   }
 }
